Transform Micro::Case::Base into Micro::Case and replace Flow::Safe by Safe::Flow
