+++ lib/micro/case/safe.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
 module Micro
-  module Case
-    class Safe < Case::Base
+  class Case
+    class Safe < ::Micro::Case
       def call
         super
       rescue => exception
         raise exception if Error::ByWrongUsage.check(exception)
+
         Failure(exception)
       end
     end
