+++ lib/micro/case/strict.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 module Micro
-  module Case
-    class Strict < Case::Base
+  class Case
+    class Strict < ::Micro::Case
       include Micro::Attributes::Features::StrictInitialize
 
-      class Safe < Case::Safe
+      class Safe < ::Micro::Case::Safe
         include Micro::Attributes::Features::StrictInitialize
       end
     end
