+++ lib/micro/case/version.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Micro
-  module Case
+  class Case
     VERSION = '1.1.0'.freeze
   end
 end
